+++ UnitTests/BraintreePayPalTests/BTPayPalRequest_Tests.swift
@@ -44,7 +44,7 @@ class BTPayPalRequest_Tests: XCTestCase {
         request.displayName = "Display Name"
         request.landingPageType = .login
         request.localeCode = .en_US
-        request.riskCorrelationId = "123-correlation-id"
+        request.riskCorrelationID = "123-correlation-id"
         request.merchantAccountID = "merchant-account-id"
         request.isShippingAddressEditable = true