+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Convert `BraintreePayPal` module to Swift
 * Breaking Changes
   * BraintreePayPal
+    * Rename `riskCorrelationId` to `riskCorrelationID`
     * Update `BTPayPalRequestLandingPageType` enum default case to `.none`
         * Update enum values
             * `.none` = 0